Refactor serial, small fixes (#7)
* Refactor serial communication

- Avoid using String to minimize heap usage

* Refactor variable names to camelCase for consistency

* Fix compiler warnings, add macros

* Increase serial input buffer size


* Add buffer size constant

* Small refactor for clarity

---------

+++ labs/lab-3/sensors.ino
@@ -8,9 +8,9 @@ void setupSensorData() {
   sensorData.speed = 0;
 }
 
-void setupSpeedSensor(int pin) {
+void setupSpeedSensor(unsigned int pin) {
   pinMode(pin, INPUT);
-  speedSensor = {0, micros()/1000000, 0};
+  speedSensor = {0, (int)(micros()/1000000), 0};
   attachInterrupt(digitalPinToInterrupt(pin), speedSensorISR, RISING); 
   if DEBUG {
     Serial.println("Speed sensor initialized");
